+++ data/group.py
@@ -27,79 +27,3 @@ group = {
         'Gen': gen_groups,
     }
 }
-
-# groups = [
-#     {
-#         "name": "Reconstruction",
-#         "subcategories": [
-#             {
-#                 "name": "Data",
-#                 "groups": recontruction_groups,
-#             },
-#             {
-#                 "name": "FastSim",
-#                 "groups": recontruction_groups,
-#             },
-#             {
-#                 "name": "FullSim",
-#                 "groups": recontruction_groups,
-#             },
-            
-#         ]
-#     },
-#     {
-#         "name": "HLT",
-#         "subcategories": [
-#             {
-#                 "name": "Data",
-#                 "groups": hlt_groups,
-#             },
-#             {
-#                 "name": "FullSim",
-#                 "groups": hlt_groups,
-#             },
-            
-#         ]
-#     },
-#     {
-#         "name": "PAGs",
-#         "subcategories": [
-#             {
-#                 "name": "Data",
-#                 "groups": pags_groups,
-#             },
-#             {
-#                 "name": "FastSim",
-#                 "groups": pags_groups,
-#             },
-#             {
-#                 "name": "FullSim",
-#                 "groups": pags_groups,
-#             },
-            
-#         ]
-#     },
-#     {
-#         "name": "HIN",
-#         "subcategories": [
-#             {
-#                 "name": "Data",
-#                 "groups": hin_groups,
-#             },
-#             {
-#                 "name": "FullSim",
-#                 "groups": hin_groups,
-#             },
-            
-#         ]
-#     },
-#     {
-#         "name": "GEN",
-#         "subcategories": [
-#             {
-#                 "name": "Gen",
-#                 "groups": gen_groups,
-#             },
-#         ]
-#     },
-# ]
