+++ database/index.py
@@ -4,3 +4,5 @@ def database_index_setup():
     database = get_database()
     db = database()
     db.database['campaign'].create_index('name')
+
+    # TODO: add index for report collection
