+++ models/campaign.py
@@ -1,12 +1,8 @@
 from models.report import Report
-from models.user import User
 from core import Model
 from datetime import datetime
 
-campaign_prefix = 'CMP'
-
 class Campaign(Model):
-    id: str
     name: str
     description: str
     deadline: datetime
@@ -15,11 +11,6 @@ class Campaign(Model):
     subcategories: list[str]
     reports: list[Report]
     active: bool
-
-    # def create(self):
-    #     campaign_id = f'{campaign_prefix}-{self._database.count(self._get_collection_name()) + 1}'
-    #     self.id = campaign_id
-    #     return self.save()
 
     @classmethod
     def get_by_name(cls, campaign_name):
