+++ models/user.py
@@ -1,4 +1,5 @@
 from core import Model
 
 class User(Model):
+    # TODO: add actual fields
     name: str
