--- react_frontend/src/components/CampaignCategoryManager.tsx
@@ -1,19 +1,0 @@
-import { Accordion, AccordionDetails, AccordionSummary, Box, Button, Divider, List, ListItem } from "@material-ui/core";
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import { faPlus, faCaretDown } from '@fortawesome/free-solid-svg-icons';
-import { CampaignCategoryList } from "./CampaignCategoryList";
-// import { mockCampaignCategories } from "../utils/mock";
-
-export function ReportCategoryManager() {
-  return (
-    <Box>
-      <Box>
-
-      </Box>
-      <h2>Campaign Category</h2>
-      {/* <CampaignCategoryList categories={mockCampaignCategories} selectable/>
-      <CampaignCategoryList categories={mockCampaignCategories} editable/> */}
-      {/* <CamapignCategoryList categories={mockCampaignCategories} selectable/> */}
-    </Box>
-  )
-}
