+++ react_frontend/src/components/ReportContentEditor.tsx
@@ -1,4 +1,4 @@
-import { useEffect, useMemo, useState } from 'react';
+import { useMemo } from 'react';
 import SimpleMDE from 'react-simplemde-editor';
 import "easymde/dist/easymde.min.css";
 import { Options } from 'easymde';
@@ -11,7 +11,6 @@ interface ReportContentEditorProp {
 export function ReportContentEditor (prop: ReportContentEditorProp) {
   const noSpellcheckerOptions = useMemo(() => {
     return {
-      // autofocus: true,
       spellChecker: false,
     } as Options;
   }, []);
