+++ react_frontend/src/pages/AdministratorPage.tsx
@@ -1,6 +1,5 @@
 import { Container } from "../components/Container";
 import { NavBar } from "../components/NavBar";
-import { ReportCategoryManager } from "../components/CampaignCategoryManager";
 
 export function AdministratorPage() {
   return (
@@ -8,7 +7,6 @@ export function AdministratorPage() {
       <NavBar />
       <Container>
         <h1>Administrator</h1>
-        <ReportCategoryManager/>
       </Container>
     </>
   )
