+++ react_frontend/src/pages/ReportPage.tsx
@@ -3,10 +3,9 @@ import { ReportContentViewer } from '../components/ReportContentViewer';
 import { ReportHeader } from '../components/ReportHeader';
 import { NavBar } from '../components/NavBar';
 import { Container } from '../components/Container';
-import { Box, Button } from "@material-ui/core"
+import { Box } from "@material-ui/core"
 import { Report, ReportEditorMode, ReportStatus } from '../types'
 import { useEffect, useState } from 'react';
-import { HorizontalLine } from '../components/HorizontalLine';
 import { useHistory, useLocation, useParams } from 'react-router';
 import queryString from 'querystring'
 import { reportService } from '../services';
@@ -16,8 +15,6 @@ export function ReportPage () {
   const { campaign, group }: any = useParams()
   const { search } = useLocation()
 
-  // const [title, setTitle] = useState<string>('Software Engineer');
-  // const [description, setDescription] = useState<string>('Lorem ipsum dolor sit amet consectetur adipisicing elit. Quia perferendis nulla saepe, tenetur ea cumque iste porro accusantium fugit consequuntur eaque aspernatur dolorum ipsam necessitatibus aperiam reprehenderit a, repellendus natus.')
   const [content, setContent] = useState<string>('');
   const [status, setStatus] = useState<ReportStatus>(ReportStatus.IN_PROGRESS);
   const [report, setReport] = useState<Report>();
@@ -40,7 +37,7 @@ export function ReportPage () {
         setEditingContent(response.data.content)
         setStatus(response.data.status)
       } else {
-        throw Error
+        throw Error('Internal Error')
       }
     }).catch(error => alert(error))
   }, [])
@@ -55,7 +52,7 @@ export function ReportPage () {
           console.log('success')
           setContent(editingContent)
         } else {
-          throw Error
+          throw Error('Internal Error')
         }
       }).catch(error => alert(error))
     } else {
@@ -72,21 +69,13 @@ export function ReportPage () {
           console.log('success')
           setStatus(newStatus as ReportStatus)
         } else {
-          throw Error
+          throw Error('Internal Error')
         }
       }).catch(error => alert(error))
     } else {
       alert('report not found!')
     }
   }
-
-  // const handleSaveTitle = (title: string) => {
-  //   setTitle(title)
-  // }
-
-  // const handleSaveDescription = (description: string) => {
-  //   setDescription(description)
-  // }
 
   const handleDiscard = () => {
     setEditingContent(content)
@@ -97,7 +86,6 @@ export function ReportPage () {
       <NavBar />
       <Container>
         { report && <Box>
-          {/* {report._id} */}
           <ReportHeader campaign={campaign} group={group} status={status} mode={mode} onChangeMode={setMode} handleSave={handleSave} handleDiscard={handleDiscard} handleChangeStatus={handleChangeStatus} />
           { mode === 'edit' && <ReportContentEditor content={editingContent} onChangeContent={setEditingContent} />}
           { (mode === 'view' || mode === 'readonly') && <ReportContentViewer content={content} />}
